+++ src/ray/gcs/gcs_server/BUILD.bazel
@@ -64,6 +64,32 @@ ray_cc_library(
 )
 
 ray_cc_library(
+    name = "node_manager",
+    hdrs = ["node_manager.h"],
+    srcs = ["node_manager.cc"],
+    deps = [
+        ":gcs_init_data",
+        ":gcs_table_storage",
+        "//src/ray/common:asio",
+        "//src/ray/common:id",
+        "//src/ray/util:logging",
+        "//src/ray/util:time",
+        "//src/ray/common:ray_config",
+        "//src/ray/rpc:node_manager_client",
+        "//src/ray/rpc:gcs_server",
+        "//src/ray/gcs/pubsub:gcs_pub_sub_lib",
+        "//src/ray/protobuf:gcs_service_cc_proto",
+        "//src/ray/protobuf:ray_syncer_cc_proto",
+        "//src/ray/util:event",
+        "@com_google_absl//absl/container:flat_hash_map",
+    ],
+    implementation_deps = [
+        "//src/ray/gcs:gcs_pb_util",
+        "@com_google_absl//absl/container:flat_hash_set",
+    ],
+)
+
+ray_cc_library(
     name = "gcs_usage_stats_client",
     srcs = ["usage_stats_client.cc"],
     hdrs = ["usage_stats_client.h"],
@@ -199,7 +225,6 @@ ray_cc_library(
         "gcs_actor_manager.cc",
         "gcs_actor_scheduler.cc",
         "gcs_autoscaler_state_manager.cc",
-        "gcs_node_manager.cc",
         "gcs_placement_group_mgr.cc",
         "gcs_placement_group_scheduler.cc",
         "gcs_resource_manager.cc",
@@ -209,7 +234,6 @@ ray_cc_library(
         "gcs_actor_manager.h",
         "gcs_actor_scheduler.h",
         "gcs_autoscaler_state_manager.h",
-        "gcs_node_manager.h",
         "gcs_placement_group_mgr.h",
         "gcs_placement_group_scheduler.h",
         "gcs_resource_manager.h",
@@ -221,6 +245,7 @@ ray_cc_library(
         ":gcs_init_data",
         ":gcs_job_manager",
         ":gcs_kv_manager",
+        ":node_manager",
         ":gcs_pubsub_handler",
         ":gcs_runtime_env_handler",
         ":gcs_server_io_context_policy",
