+++ src/ray/gcs/gcs_server/gcs_actor_scheduler.h
@@ -27,8 +27,8 @@
 #include "ray/common/id.h"
 #include "ray/common/scheduling/scheduling_ids.h"
 #include "ray/common/task/task_spec.h"
-#include "ray/gcs/gcs_server/gcs_node_manager.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
+#include "ray/gcs/gcs_server/node_manager.h"
 #include "ray/raylet/scheduling/cluster_task_manager.h"
 #include "ray/raylet_client/raylet_client.h"
 #include "ray/rpc/node_manager/node_manager_client.h"
