+++ src/ray/gcs/gcs_server/tests/BUILD.bazel
@@ -105,16 +105,17 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "gcs_node_manager_test",
+    name = "node_manager_test",
     size = "small",
     srcs = [
-        "gcs_node_manager_test.cc",
+        "node_manager_test.cc",
     ],
     tags = ["team:core"],
     deps = [
+        "//:ray_fakes",
         ":gcs_server_test_util",
         "//src/mock/ray/pubsub:mock_publisher",
-        "//src/ray/gcs/gcs_server:gcs_server_lib",
+        "//src/ray/gcs/gcs_server:node_manager",
         "//src/ray/gcs/tests:gcs_test_util_lib",
         "@com_google_googletest//:gtest_main",
     ],
@@ -396,19 +397,20 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "gcs_node_manager_export_event_test",
+    name = "node_manager_export_event_test",
     size = "small",
-    srcs = ["export_api/gcs_node_manager_export_event_test.cc"],
+    srcs = ["export_api/node_manager_export_event_test.cc"],
     tags = [
         "no_windows",
         "team:core",
     ],
     deps = [
-        ":gcs_server_test_util",
+        "//:ray_fakes",
         "//src/mock/ray/pubsub:mock_publisher",
-        "//src/ray/gcs/gcs_server:gcs_server_lib",
+        "//src/ray/util:string_utils",
+        "//src/ray/gcs/gcs_server:node_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
         "//src/ray/gcs/tests:gcs_test_util_lib",
-        "@com_google_googletest//:gtest_main",
+        "@com_google_googletest//:gtest",
     ],
 )
