+++ src/ray/gcs/gcs_server/tests/export_api/gcs_actor_manager_export_event_test.cc
@@ -28,7 +28,7 @@
 #include "ray/gcs/gcs_server/gcs_kv_manager.h"
 #include "ray/gcs/store_client/in_memory_store_client.h"
 #include "mock/ray/gcs/gcs_server/gcs_kv_manager.h"
-#include "mock/ray/gcs/gcs_server/gcs_node_manager.h"
+#include "mock/ray/gcs/gcs_server/node_manager.h"
 #include "ray/pubsub/publisher.h"
 #include "ray/util/event.h"
 // clang-format on
