+++ src/ray/gcs/gcs_server/tests/export_api/node_manager_export_event_test.cc
@@ -20,17 +20,13 @@
 #include <thread>
 #include <vector>
 
-#include "ray/gcs/gcs_server/tests/gcs_server_test_util.h"
+#include "fakes/ray/rpc/raylet/raylet_client.h"
+#include "mock/ray/pubsub/publisher.h"
+#include "ray/gcs/gcs_server/node_manager.h"
 #include "ray/gcs/store_client/in_memory_store_client.h"
 #include "ray/gcs/tests/gcs_test_util.h"
 #include "ray/util/event.h"
 #include "ray/util/string_utils.h"
-
-// clang-format off
-#include "ray/rpc/node_manager/node_manager_client.h"
-#include "ray/rpc/node_manager/raylet_client_pool.h"
-#include "mock/ray/pubsub/publisher.h"
-// clang-format on
 
 using json = nlohmann::json;
 
@@ -46,9 +42,11 @@ std::string GenerateLogDir() {
 class GcsNodeManagerExportAPITest : public ::testing::Test {
  public:
   GcsNodeManagerExportAPITest() {
-    raylet_client_ = std::make_shared<GcsServerMocker::MockRayletClient>();
+    auto raylet_client = std::make_shared<FakeRayletClient>();
     client_pool_ = std::make_unique<rpc::RayletClientPool>(
-        [this](const rpc::Address &) { return raylet_client_; });
+        [raylet_client = std::move(raylet_client)](const rpc::Address &) {
+          return raylet_client;
+        });
     gcs_publisher_ = std::make_unique<gcs::GcsPublisher>(
         std::make_unique<ray::pubsub::MockPublisher>());
     gcs_table_storage_ = std::make_unique<gcs::GcsTableStorage>(
@@ -78,7 +76,6 @@ class GcsNodeManagerExportAPITest : public ::testing::Test {
 
  protected:
   std::unique_ptr<gcs::GcsTableStorage> gcs_table_storage_;
-  std::shared_ptr<GcsServerMocker::MockRayletClient> raylet_client_;
   std::unique_ptr<rpc::RayletClientPool> client_pool_;
   std::shared_ptr<gcs::GcsPublisher> gcs_publisher_;
   instrumented_io_context io_service_;
