+++ src/ray/gcs/gcs_server/tests/gcs_actor_manager_test.cc
@@ -27,7 +27,7 @@
 #include "ray/gcs/store_client/in_memory_store_client.h"
 #include "ray/pubsub/publisher.h"
 #include "mock/ray/gcs/gcs_server/gcs_kv_manager.h"
-#include "mock/ray/gcs/gcs_server/gcs_node_manager.h"
+#include "mock/ray/gcs/gcs_server/node_manager.h"
 // clang-format on
 
 namespace ray {
