+++ src/ray/gcs/gcs_server/tests/gcs_autoscaler_state_manager_test.cc
@@ -25,8 +25,8 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "mock/ray/gcs/gcs_server/gcs_actor_manager.h"
-#include "mock/ray/gcs/gcs_server/gcs_node_manager.h"
 #include "mock/ray/gcs/gcs_server/gcs_placement_group_mgr.h"
+#include "mock/ray/gcs/gcs_server/node_manager.h"
 #include "mock/ray/gcs/store_client/store_client.h"
 #include "mock/ray/rpc/worker/core_worker_client.h"
 #include "ray/common/asio/instrumented_io_context.h"
