+++ src/ray/gcs/gcs_server/tests/gcs_resource_manager_test.cc
@@ -19,7 +19,7 @@
 #include <string>
 
 #include "gtest/gtest.h"
-#include "mock/ray/gcs/gcs_server/gcs_node_manager.h"
+#include "mock/ray/gcs/gcs_server/node_manager.h"
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/gcs/tests/gcs_test_util.h"
 #include "ray/raylet/scheduling/cluster_resource_manager.h"
